Create addTwoNumbers_v2.py
This is a cleaner version of my original addTwoNumbers program. The only difference is I cleaned the code up a little bit. The reason I'm not overwriting the original file is because the original file was 33% faster so I'd like to be able to reference both files in the future